build: fix NCCL 2.30 RAS timeout helper types
NCCL 2.30 tightened the RAS timeout arithmetic enough that RCCL's plain long long CLOCK_UNITS_PER_SEC macro now collides with int64_t wakeup state inside std::min call sites. Cast the shared constant to int64_t and add the missing <algorithm> includes where the new std::min usage appears.

Test: build proceeds past the rasnet.cc and collectives.cc std::min overload failures.

Verified: rm -f build/hipify/src/ras/rasnet.cc build/hipify/src/ras/collectives.cc build/hipify/src/ras/ras_internal.h && timeout 1500 make -j$(nproc) advances to the remaining GIN/net/plugin errors.

+++ projects/rccl/src/ras/ras_internal.h
@@ -191,7 +191,7 @@ static inline size_t rasMsgLength(rasMsgType type, rasCollectiveType collType = 
 #define RAS_INCREMENT 4
 
 // Our clock has nanosecond resolution.
-#define CLOCK_UNITS_PER_SEC (1000000000LL)
+#define CLOCK_UNITS_PER_SEC ((int64_t)1000000000LL)
 
 // Keep-alive messages are sent no sooner than a second after the last message was sent down a particular connection.
 #define RAS_KEEPALIVE_INTERVAL (1*CLOCK_UNITS_PER_SEC*ncclParamRasTimeoutFactor())
